Replaced isMatchedFile and fix for hardcoded values
